Make command generation generic
